feat: リファクタリング - 冗長なタイピングコンポーネントを統合
- 統合TypingGameコンポーネントを作成し、3つの個別ページから共通ロジックを抽出
- useTypingDataコンポーザブルでAPI呼び出しを統一
- 冗長なファイルを削除（skilltest*.vue, useApiFetch.ts, word.ts）
- TypeScript strict モードを有効化
- コード量を約60%削減（1800行 → 700行程度）
