--- src/composables/useApiFetch.ts
@@ -1,6 +1,0 @@
-export const useApiFetch = () => {
-  const apiFetch = $fetch.create({
-    baseURL: 'http://localhost:4000',
-  })
-  return { apiFetch }
-}
