--- src/pages/skilltest3.vue
@@ -1,16 +1,0 @@
-<template>
-  <h1 id="title">
-    腕試しをしてみよう！
-  </h1>
-  <hr />
-  <br /><br />
-  <div id="parent">
-    <p>応用練習</p>
-  </div>
-  <br />
-  <br />
-  <div id="start">
-    <!--タイピング画面に遷移 -->
-  </div>
-  <br />
-</template>
