--- src/pages/skilltest4.vue
@@ -1,16 +1,0 @@
-<template>
-  <h1 id="title">
-    腕試しをしてみよう！
-  </h1>
-  <hr />
-  <br /><br />
-  <div id="parent">
-    <p>有名文学作品冒頭タイピング</p>
-  </div>
-  <br />
-  <br />
-  <div id="start">
-    <!--タイピング画面に遷移 -->
-  </div>
-  <br />
-</template>
